[Messenger] Rename Adapters to Transports
