+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -971,7 +971,7 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('messenger')
                     ->info('Messenger configuration')
                     ->{!class_exists(FullStack::class) && interface_exists(MessageBusInterface::class) ? 'canBeDisabled' : 'canBeEnabled'}()
-                    ->fixXmlConfig('adapter')
+                    ->fixXmlConfig('transport')
                     ->fixXmlConfig('bus', 'buses')
                     ->children()
                         ->arrayNode('routing')
@@ -1024,7 +1024,7 @@ class Configuration implements ConfigurationInterface
                         ->end()
                         ->scalarNode('encoder')->defaultValue('messenger.transport.serializer')->end()
                         ->scalarNode('decoder')->defaultValue('messenger.transport.serializer')->end()
-                        ->arrayNode('adapters')
+                        ->arrayNode('transports')
                             ->useAttributeAsKey('name')
                             ->arrayPrototype()
                                 ->beforeNormalization()
