+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_transports.php
@@ -3,7 +3,7 @@
 $container->loadFromExtension('framework', array(
     'serializer' => true,
     'messenger' => array(
-        'adapters' => array(
+        'transports' => array(
             'default' => 'amqp://localhost/%2f/messages',
             'customised' => array(
                 'dsn' => 'amqp://localhost/%2f/messages?exchange_name=exchange_name',
