+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/messenger_transport_no_serializer.yml
@@ -4,5 +4,5 @@ framework:
     messenger:
         serializer:
             enabled: true
-        adapters:
+        transports:
             default: 'amqp://localhost/%2f/messages'
