+++ src/Symfony/Component/Messenger/Handler/Locator/ContainerHandlerLocator.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Messenger;
+namespace Symfony\Component\Messenger\Handler\Locator;
 
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
