+++ src/Symfony/Component/Messenger/Transport/AmqpExt/Connection.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Messenger\Adapter\AmqpExt;
+namespace Symfony\Component\Messenger\Transport\AmqpExt;
 
 /**
  * An AMQP connection.
